requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-google-genai
google-generativeai
youtube-transcript-api
faiss-cpu
python-dotenv
